chapter9/backend/requirements.txt
anyio==3.6.1
click==8.1.3
colorama==0.4.5
dnslib==0.9.20
dnspython==2.2.1
fastapi==0.79.0
h11==0.13.0
idna==3.3
motor==3.0.0
numpy==1.23.1
pandas==1.4.3
pydantic==1.9.1
pymongo==4.2.0
python-dateutil==2.8.2
python-decouple==3.6
python-http-client==3.3.7
PythonDNS==0.1
pytz==2022.1
sendgrid==6.9.7
six==1.16.0
sniffio==1.2.0
starkbank-ecdsa==2.0.3
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
